# api-gateway/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Second Opinion Platform API Gateway
  description: |
    API Gateway for the Second Opinion Platform microservices architecture.
    
    This gateway provides unified access to all platform services including:
    - Patient Identity Service (Authentication & Authorization)
    - Case Management Service (Medical cases & documents)
    - AI Analysis Service (Medical AI insights)
    - Professional Service (Medical professionals)
    - Notification Service (Multi-channel communications)
    
    ## Authentication
    Most endpoints require JWT Bearer token authentication obtained from the identity service.
    
    ## Rate Limiting
    API endpoints are rate limited based on user type and endpoint category.
    
    ## Error Handling
    All responses follow a consistent error format with appropriate HTTP status codes.
    
  version: 1.0.0
  contact:
    name: Second Opinion Platform Team
    email: support@secondopinion.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.secondopinion.com
    description: Production server
paths:
...
  /api/v1/auth/login:
    post:
      summary: User login
      description: Authenticate user and return JWT token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid request
        '401':
          description: Invalid credentials
        '429':
          description: Too many login attempts
  /api/v1/auth/register:
    post:
      summary: User registration
      description: Register a new user account
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
      responses:
        '201':
          description: Registration successful
        '400':
          description: Invalid request
        '409':
          description: Email already exists

  # Case Management Routes
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        accessToken:
          type: string
        user:
          type: object
          properties:
            userId:
              type: string
            email:
              type: string
            role:
              type: string
    UserRegistration:
      type: object
      required:
        - email
        - password
        - firstName
        - lastName
        - dateOfBirth
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date
        phone:
          type: string
        preferredChannel:
          type: string
          enum: [EMAIL, SMS, PHONE]
...